Add BaseResourceChangeSet mixin and source metadata identifier validators
